// src/app/shared/enums/chips.enum.ts
export enum CHIPS {
  WINNER = 'winner',
  LOSER = 'serial loser',
  HIGH_DMG = 'crazy damage',
  LOW_DMG = 'child\'s damage',
  HIGH_VISION = 'good vision',
  LOW_VISION = 'bad vision',
  HIGH_CS = 'minion eater',
  LOW_CS = 'minion hater',
  HIGH_DEATHS = 'the feeder',
  LOW_DEATHS = 'the survivor',
  HIGH_ASSISTS = 'the support',
  LOW_ASSISTS = 'help us!',
  HIGH_KILL_COUNT = 'the monster',
  LOW_KILL_COUNT = 'be more aggressive'
}